Updated Template
uses latex conditional, rather than etoolbox

+++ McrRaspJam/013_Objects/2_object/Vector3.java
@@ -2,15 +2,15 @@ import java.lang.*;
 
 public class Vector3
 {
-	public float x;
-	public float y;
-	public float z;
+	public double x;
+	public double y;
+	public double z;
 	
-	public float magnitude()
+	public double magnitude()
 	{
 		// Magnitude is the 
 		// square root of x^2 + y^2 + z^2
-		float sum = Math.pow(x, 2) + Math.pow(y, 2) + Math.pow(z, 2);
+		double sum = Math.pow(x, 2) + Math.pow(y, 2) + Math.pow(z, 2);
 		
 		return Math.sqrt(sum);
 	}
